Day 5 part 2
Finally got around to this. It's more fiddly, than anything, and I managed
to hit the right answer on first attempt (ignoring getting the source and
dest reversed in the input parsing).

I really should write unit tests for `map_seed_range`, but to be honest
I'm so over this problem, and it's not like this code will ever be used
again (ha ha ha)